Add explicit demos.sandbox packages to sandbox app classes
