+++ demos/sandbox-appyx-navigation/android/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Appyx">
         <activity
-            android:name="com.bumble.appyx.navigation.MainActivity"
+            android:name=".MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Appyx.Starting">
             <intent-filter>
