--- demos/sandbox-appyx-navigation/android/src/main/kotlin/com/bumble/appyx/navigation/MainActivity.kt
@@ -1,62 +1,0 @@
-package com.bumble.appyx.navigation
-
-import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.compose.animation.ExperimentalAnimationApi
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.ExperimentalUnitApi
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.bumble.appyx.navigation.integration.NodeActivity
-import com.bumble.appyx.navigation.integration.NodeHost
-import com.bumble.appyx.navigation.modality.NodeContext
-import com.bumble.appyx.navigation.node.container.MainNavNode
-import com.bumble.appyx.navigation.platform.AndroidLifecycle
-import com.bumble.appyx.navigation.ui.AppyxSampleAppTheme
-
-@ExperimentalUnitApi
-@ExperimentalAnimationApi
-@ExperimentalComposeUiApi
-class MainActivity : NodeActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        installSplashScreen()
-        super.onCreate(savedInstanceState)
-        setContent {
-            AppyxSampleAppTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(color = MaterialTheme.colorScheme.background) {
-                    NodeHost(
-                        lifecycle = AndroidLifecycle(LocalLifecycleOwner.current.lifecycle),
-                        integrationPoint = appyxV2IntegrationPoint,
-                    ) {
-                        MainNavNode(
-                            nodeContext = it,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-
-@Preview(showBackground = true)
-@Composable
-@ExperimentalUnitApi
-@ExperimentalAnimationApi
-@ExperimentalComposeUiApi
-fun DefaultPreview() {
-    AppyxSampleAppTheme {
-        Column {
-            MainNavNode(
-                nodeContext = NodeContext.root(null),
-            ).Compose()
-        }
-    }
-}
