+++ demos/sandbox-appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/demos/sandbox/navigation/MainNavItem.kt
@@ -1,0 +1,55 @@
+package com.bumble.appyx.demos.sandbox.navigation
+
+import androidx.compose.material.icons.Icons.Filled
+import androidx.compose.material.icons.Icons.Outlined
+import androidx.compose.material.icons.filled.ViewCarousel
+import androidx.compose.material.icons.filled.WebStories
+import androidx.compose.material.icons.outlined.GridView
+import androidx.compose.material.icons.outlined.ViewCarousel
+import androidx.compose.material.icons.outlined.WebStories
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.rotate
+import com.bumble.appyx.demos.sandbox.navigation.node.backstack.BackStackExamplesNode
+import com.bumble.appyx.demos.sandbox.navigation.node.promoter.PromoterNode
+import com.bumble.appyx.demos.sandbox.navigation.node.spotlight.SpotlightNode
+import com.bumble.appyx.demos.sandbox.navigation.ui.icons.GridViewCustom
+import com.bumble.appyx.utils.material3.AppyxNavItem
+import com.bumble.appyx.utils.multiplatform.Parcelable
+import com.bumble.appyx.utils.multiplatform.Parcelize
+
+@Parcelize
+enum class MainNavItem : Parcelable {
+    BACKSTACK, SPOTLIGHT, PROMOTER;
+
+    companion object {
+        val resolver: (com.bumble.appyx.demos.sandbox.navigation.MainNavItem) -> AppyxNavItem = { navBarItem ->
+            when (navBarItem) {
+                com.bumble.appyx.demos.sandbox.navigation.MainNavItem.BACKSTACK -> AppyxNavItem(
+                    text = "Back stack",
+                    unselectedIcon = Outlined.WebStories,
+                    selectedIcon = Filled.WebStories,
+                    modifier = Modifier.rotate(180f),
+                    node = { BackStackExamplesNode(it) }
+                )
+
+                com.bumble.appyx.demos.sandbox.navigation.MainNavItem.SPOTLIGHT -> AppyxNavItem(
+                    text = "Spotlight",
+                    unselectedIcon = Outlined.ViewCarousel,
+                    selectedIcon = Filled.ViewCarousel,
+                    node = { SpotlightNode(it) }
+                )
+
+                com.bumble.appyx.demos.sandbox.navigation.MainNavItem.PROMOTER -> AppyxNavItem(
+                    text = "Promoter",
+                    unselectedIcon = Outlined.GridView,
+                    selectedIcon = Filled.GridViewCustom,
+                    modifier = Modifier.rotate(45f),
+                    node = { PromoterNode(it) }
+                )
+            }
+        }
+    }
+}
+
+
+
