+++ demos/sandbox-appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/demos/sandbox/navigation/ui/Shape.kt
@@ -1,0 +1,12 @@
+package com.bumble.appyx.demos.sandbox.navigation.ui
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Shapes
+import androidx.compose.ui.unit.dp
+
+
+val shapes = Shapes(
+    small = RoundedCornerShape(4.dp),
+    medium = RoundedCornerShape(16.dp),
+    large = RoundedCornerShape(0.dp)
+)
