+++ demos/sandbox-appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/demos/sandbox/navigation/ui/TextButton.kt
@@ -1,0 +1,20 @@
+package com.bumble.appyx.demos.sandbox.navigation.ui
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun TextButton(
+    text: String,
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit
+) {
+    Button(modifier = modifier, shape = RoundedCornerShape(6.dp), onClick = onClick) {
+        Text(textAlign = TextAlign.Center, text = text)
+    }
+}
