--- demos/sandbox-appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/Ext.kt
@@ -1,35 +1,0 @@
-package com.bumble.appyx.navigation
-
-import com.bumble.appyx.navigation.ui.atomic_tangerine
-import com.bumble.appyx.navigation.ui.manatee
-import com.bumble.appyx.navigation.ui.md_amber_500
-import com.bumble.appyx.navigation.ui.md_blue_500
-import com.bumble.appyx.navigation.ui.md_blue_grey_500
-import com.bumble.appyx.navigation.ui.md_cyan_500
-import com.bumble.appyx.navigation.ui.md_grey_500
-import com.bumble.appyx.navigation.ui.md_indigo_500
-import com.bumble.appyx.navigation.ui.md_light_blue_500
-import com.bumble.appyx.navigation.ui.md_light_green_500
-import com.bumble.appyx.navigation.ui.md_lime_500
-import com.bumble.appyx.navigation.ui.md_pink_500
-import com.bumble.appyx.navigation.ui.md_teal_500
-import com.bumble.appyx.navigation.ui.silver_sand
-import com.bumble.appyx.navigation.ui.sizzling_red
-
-val colors = listOf(
-    manatee,
-    sizzling_red,
-    atomic_tangerine,
-    silver_sand,
-    md_pink_500,
-    md_indigo_500,
-    md_blue_500,
-    md_light_blue_500,
-    md_cyan_500,
-    md_teal_500,
-    md_light_green_500,
-    md_lime_500,
-    md_amber_500,
-    md_grey_500,
-    md_blue_grey_500
-)
