--- demos/sandbox-appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/Savers.kt
@@ -1,20 +1,0 @@
-package com.bumble.appyx.navigation
-
-import androidx.compose.runtime.saveable.mapSaver
-import androidx.compose.ui.graphics.Color
-
-val ColorSaver = run {
-    val redKey = "Red"
-    val greenKey = "Green"
-    val blueKey = "Blue"
-    mapSaver(
-        save = { mapOf(redKey to it.red, greenKey to it.green, blueKey to it.blue) },
-        restore = {
-            Color(
-                red = it[redKey] as Float,
-                green = it[greenKey] as Float,
-                blue = it[blueKey] as Float
-            )
-        }
-    )
-}
