--- demos/sandbox-appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/ui/Color.kt
@@ -1,23 +1,0 @@
-package com.bumble.appyx.navigation.ui
-
-import androidx.compose.ui.graphics.Color
-
-val purple200 = Color(0xFFBB86FC)
-val purple500 = Color(0xFF6200EE)
-val purple700 = Color(0xFF3700B3)
-val teal200 = Color(0xFF03DAC5)
-
-val space_cadet = Color(0xFF2B2D42)
-val black_coral = Color(0xFF5C6378)
-val manatee = Color(0xFF8D99AE)
-val silver_sand = Color(0xFFBDC6D1)
-val alice_blue = Color(0xFFEDF2F4)
-val sizzling_red = Color(0xFFF05D5E)
-val imperial_red = Color(0xFFEF233C)
-val amaranth_red = Color(0xFFD90429)
-val atomic_tangerine = Color(0xFFF0965D)
-
-val appyx_yellow1 = Color(0xFFFFC629)
-val appyx_yellow2 = Color(0xFFFFE54A)
-val appyx_dark = Color(0xFF1F2126)
-val appyx_bright = Color(0xFFFFFFFF)
