--- demos/sandbox-appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/ui/icons/GridViewCustom.kt
@@ -1,71 +1,0 @@
-package com.bumble.appyx.navigation.ui.icons
-
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.materialIcon
-import androidx.compose.material.icons.materialPath
-import androidx.compose.ui.graphics.vector.ImageVector
-
-
-
-internal val Icons.Filled.GridViewCustom: ImageVector
-    // Based on Icons.Filled.GridView
-    get() {
-        if (_gridView != null) {
-            return _gridView!!
-        }
-        _gridView = materialIcon(name = "Outlined.GridView") {
-            materialPath {
-                moveTo(3.0f, 3.0f)
-                verticalLineToRelative(8.0f)
-                horizontalLineToRelative(8.0f)
-                verticalLineTo(3.0f)
-                horizontalLineTo(3.0f)
-                close()
-//                moveTo(9.0f, 9.0f)
-//                horizontalLineTo(5.0f)
-//                verticalLineTo(5.0f)
-//                horizontalLineToRelative(4.0f)
-//                verticalLineTo(9.0f)
-//                close()
-                moveTo(3.0f, 13.0f)
-                verticalLineToRelative(8.0f)
-                horizontalLineToRelative(8.0f)
-                verticalLineToRelative(-8.0f)
-                horizontalLineTo(3.0f)
-                close()
-                moveTo(9.0f, 19.0f)
-                horizontalLineTo(5.0f)
-                verticalLineToRelative(-4.0f)
-                horizontalLineToRelative(4.0f)
-                verticalLineTo(19.0f)
-                close()
-                moveTo(13.0f, 3.0f)
-                verticalLineToRelative(8.0f)
-                horizontalLineToRelative(8.0f)
-                verticalLineTo(3.0f)
-                horizontalLineTo(13.0f)
-                close()
-                moveTo(19.0f, 9.0f)
-                horizontalLineToRelative(-4.0f)
-                verticalLineTo(5.0f)
-                horizontalLineToRelative(4.0f)
-                verticalLineTo(9.0f)
-                close()
-                moveTo(13.0f, 13.0f)
-                verticalLineToRelative(8.0f)
-                horizontalLineToRelative(8.0f)
-                verticalLineToRelative(-8.0f)
-                horizontalLineTo(13.0f)
-                close()
-                moveTo(19.0f, 19.0f)
-                horizontalLineToRelative(-4.0f)
-                verticalLineToRelative(-4.0f)
-                horizontalLineToRelative(4.0f)
-                verticalLineTo(19.0f)
-                close()
-            }
-        }
-        return _gridView!!
-    }
-
-private var _gridView: ImageVector? = null
