+++ demos/sandbox-appyx-navigation/desktop/build.gradle.kts
@@ -32,7 +32,7 @@ kotlin {
 
 compose.desktop {
     application {
-        mainClass = "com.bumble.appyx.navigation.MainKt"
+        mainClass = "com.bumble.appyx.demos.sandbox.navigation.MainKt"
         nativeDistributions {
             targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
             packageName = "AppyxNavigationDesktop"
